requirements.txt
category_encoders==2.6.1
fastapi==0.100.0
numpy==1.21.5
pandas==1.4.4
pydantic==1.10.8
pytest==7.1.2
python-dotenv==1.0.0
scikit_learn==1.0.2
starlette==0.27.0
uvicorn==0.22.0
